requirements.txt
# ---------- FastAPI ----------
fastapi==0.115.6
uvicorn==0.30.6
python-multipart==0.0.9

# ---------- PDF Processing ----------
pdfplumber==0.11.4
camelot-py==0.11.0
PyMuPDF==1.24.9

# ---------- OCR ----------
pytesseract==0.3.13
pillow==10.4.0
opencv-python==4.10.0.84

# ---------- Data ----------
pandas==2.2.2
numpy==1.26.4

# ---------- Utilities ----------
langdetect==1.0.9
python-dotenv==1.0.0

# ---------- LLM ----------
groq==0.4.1
httpx==0.27.2  # pin to avoid proxies arg error with Groq client

# ---------- Embeddings + Vector DB ----------
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1

# ---------- Document Processing ----------
python-docx
openpyxl

